image caption/requirements.txt
flask
numpy
pandas
tensorflow
keras
pillow
nltk
